requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
crispy-bootstrap5==0.6
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.5
django-braces==1.15.0
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-rest-framework==0.1.0
django-star-ratings==0.9.2
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
kombu==5.2.4
Pillow==9.2.0
pipenv==2022.6.7
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2022.1
vine==5.0.0
virtualenv==20.15.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.2.0
